portfolio_service/requirements.txt
flask
flask-cors
flask-sqlalchemy
requests
pyjwt